Added a video player
The video player is a game object, so should be able to be moved using Ryan's window moving code. Should also be able to be resized in the same way (although this will stretch the video).

Currently uses a video stored within the assets folder. It seems like you can make it work using video links so will try and use that instead.